plugins/pencheff/pyproject.toml
[project]
name = "pencheff"
version = "0.1.0"
description = "AI Penetration Testing Agent - MCP Server for Claude Code"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "pyjwt>=2.0.0",
    "cryptography>=40.0.0",
    "jinja2>=3.0.0",
    "pyyaml>=6.0",
    "dnspython>=2.4.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "anyio>=4.0.0",
]

[project.optional-dependencies]
full = [
    "python-nmap>=0.7.1",
    "boto3>=1.28.0",
    "paramiko>=3.0.0",
]

[project.scripts]
pencheff = "pencheff.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["pencheff"]
